feat: refactor authentication logic to use SessionProvider and update related tests
